chat-bot-with-memory/requirements.txt
python-dotenv==1.0.1
langchain_openai==0.1.14
openai==1.68.2
langchain==0.2.17
